Poorva | code cleanup
