+++ cli/src/main/kotlin/norm/api/NormApi.kt
@@ -13,13 +13,11 @@ class NormApi(
         connection: Connection
 ) {
     private val sqlAnalyzer = SqlAnalyzer(connection)
-    private val codeGenerator = CodeGenerator()
-
     /**
      * Generates code of classes against given SQL Query
      */
     fun generate(query: String, packageName: String, baseName: String): String {
         val sqlModel = sqlAnalyzer.sqlModel(query)
-        return codeGenerator.generate(listOf(sqlModel), packageName, baseName)
+        return CodeGenerator.generate(listOf(sqlModel), packageName, baseName)
     }
 }
