+++ cli/src/test/kotlin/norm/test/utils/TestUtils.kt
@@ -11,7 +11,7 @@ fun toArgs(str: String): Array<String> = str.split(" ").toTypedArray()
 
 fun codegen(conn: Connection, query: String, pkg: String, base: String): String {
     val sqlModel = SqlAnalyzer(conn).sqlModel(query)
-    return CodeGenerator().generate(listOf(sqlModel), pkg, base).trim()
+    return CodeGenerator.generate(listOf(sqlModel), pkg, base).trim()
 }
 
 fun String.readAsResource(): String = object {}.javaClass.getResource(this).readText().trim()
